Extract device-proxy to a feature
Change-type: patch
