+++ src/features/device-proxy/index.ts
@@ -1,0 +1,8 @@
+import type { Application } from 'express';
+
+import { apiKeyMiddleware } from '../../infra/auth';
+import { proxy } from './device-proxy';
+
+export const setup = (app: Application) => {
+	app.post(/^\/supervisor(\/.+)$/, apiKeyMiddleware, proxy);
+};
